requirements.txt
Flask
pymongo
psutil
Flask-Login
python-dotenv
werkzeug
dnspython